Problems renaming the GMSH physical groups
Renaming the GMSH physical groups to match the GalerkinToolkit
nodal ordering causes problems in the permutation function.
This is because the opposite faces no longer match up. The
GalerkinToolkit nodal ordering is an "inverse Z" ordering starting
from the bottom left, while the GMSH ordering begins with vertices,
then curves, then surfaces.

TODO:
[] Regarding experiments, since the baseline mesh generator is GMSH,
need to write a function that creates a geometry matching that of the
final cartesian geometry of GalerkinToolkit, and then have GMSH
subsequently mesh that geometry (timing this process).