Not possible to check node ownership on interface?
To assert which parts own the correct nodes, this would require that a
given mesh part has some methods like `boundary(mesh, 0)` to get the
node indices corresponding to the boundary, and a subset of the boundary
nodes would be on the interface (i.e., for part 1, the interface is part
3 to the north, part 2 to the east, and part 4 is the intersection
of all the parts). I don't think we have a data structure that has the
boundary information for the final cells in the coarse mesh, or if we do,
that boundary information would likely only have fine node ids, and what
we need is final node ids since the
`gk.node_indices(ids)` probably deals with the final indices? Need to look into
this later, and just move on to PMesh tests for all desired geometries